move refresh to keygen
