+++ protocols/cmp/keygen/base_test.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"fmt"
