+++ protocols/cmp/keygen/fake.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"crypto/rand"
