+++ protocols/cmp/keygen/public_test.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"crypto/rand"
