+++ protocols/cmp/keygen/result.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 type Result struct {
 	Session *Session
