+++ protocols/cmp/keygen/round1.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"crypto/rand"
