+++ protocols/cmp/keygen/round3.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"bytes"
