+++ protocols/cmp/keygen/round4.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"errors"
