+++ protocols/cmp/keygen/session.go
@@ -1,4 +1,4 @@
-package refresh
+package keygen
 
 import (
 	"bytes"
