+++ protocols/cmp/sign/local_party.go
@@ -3,7 +3,6 @@ package sign
 import (
 	"github.com/taurusgroup/cmp-ecdsa/pkg/math/curve"
 	"github.com/taurusgroup/cmp-ecdsa/pkg/paillier"
-	refresh2 "github.com/taurusgroup/cmp-ecdsa/protocols/cmp/refresh"
 )
 
 // LocalParty is the state we store for a remote party.
